requirements.txt
azure-core==1.7.0
azure-eventhub==5.1.0
azure-iot-device==2.1.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
janus==0.4.0
paho-mqtt==1.5.0
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-unixsocket==0.2.0
six==1.15.0
uamqp==1.2.9
urllib3==1.25.9
